Fixes from FlixelCommunity

+++ flixel-core/src/org/flixel/system/FlxQuadTree.java
@@ -309,13 +309,17 @@ public class FlxQuadTree extends FlxRect
 	 */
 	public void destroy()
 	{
-		_headA.destroy();
+		if(_headA != null)
+			_headA.destroy();
 		_headA = null;
-		//_tailA.destroy();
+		//if(_tailA != null)
+		//	_tailA.destroy();
 		_tailA = null;
-		_headB.destroy();
+		if(_headB != null)
+			_headB.destroy();
 		_headB = null;
-		//_tailB.destroy();
+		//if(_tailB != null)
+			//_tailB.destroy();
 		_tailB = null;
 		
 		if(_northWestTree != null)
